+++ packages/common/src/index.ts
@@ -14,4 +14,5 @@ export * from "./trees/InMemoryMerkleTreeStorage";
 export * from "./trees/RollupMerkleTree";
 export * from "./events/EventEmitterProxy";
 export * from "./trees/MockAsyncMerkleStore";
-export * from "./ProvableMethodExecutionContext";
+export * from "./zkProgram/ProvableMethodExecutionContext";
+export * from "./zkProgram/WithZkProgram";
